pyproject.toml
[tool.poetry]
name = "fonts-dataset"
version = "0.1.0"
description = ""
authors = ["Yasuaki Uechi <y@uechi.io>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
Pillow = "^6.0"
fonttools = "^3.41"
protobuf = "^3.7"
numpy = "^1.16"
click = "^7.0"

[tool.poetry.dev-dependencies]
pytest = "^4.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
